+++ src/components/horizontal-bar-chart.stories.tsx
@@ -1,5 +1,9 @@
 import data from "mocks/sunshine-powerStability-426.json";
+import { useState } from "react";
 
+import { useQueryStatePowerStabilityCardFilters } from "src/domain/query-states";
+
+import { PowerStabilityCard } from "./power-stability-card";
 import { PowerStabilityChart } from "./power-stability-chart";
 import { DesignGrid, DesignStory } from "./storybook/base-style";
 
@@ -20,6 +24,41 @@ export const HorizontalBarChart = () => {
           compareWith={[]}
         />
       </DesignGrid>
+    </DesignStory>
+  );
+};
+
+export const HorizontalBarChartCard = () => {
+  const [state, setQueryState] = useState<
+    ReturnType<typeof useQueryStatePowerStabilityCardFilters>[0]
+  >({
+    compareWith: [],
+    viewBy: "latest",
+    duration: "total",
+    overallOrRatio: "overall",
+  });
+
+  return (
+    <DesignStory
+      title="Power Stability Card"
+      reference="ElCom Library (Sunshine Indicators)"
+    >
+      <PowerStabilityCard
+        state={state}
+        setQueryState={(newState) =>
+          newState && setQueryState((prev) => ({ ...prev, ...newState }))
+        }
+        peerGroup={{
+          energyDensity: "na",
+          settlementDensity: "unknown",
+        }}
+        updateDate={data.saidi.yearlyData[0].year.toString()}
+        latestYear={2024}
+        operatorId="11"
+        operatorLabel="Elektrizitätswerk des Kantons Schaffhausen AG"
+        observations={data.saidi.yearlyData.filter((x) => x.year === 2024)}
+        cardTitle="Average Power Outage Frequency (SAIFI)"
+      />
     </DesignStory>
   );
 };